surya: Update custom `libcameraservice.so`
Due to a bug in the device camera, timestamp cannot be acquired during
video recording. this leads to very strange videos which are hours long
but only the last few seconds actually contains any video.

To fix this bug a custom libcameraservice had to be adopted.
This code modification is detailed in the device tree commit.

Signed-off-by: Ray <vincentrayananda@gmail.com>
Change-Id: I7e323d53498a8bb26c8eea95a38357378e5c1249
Signed-off-by: Genchler <ben.genchler@gmail.com>